Issue #1792 Refactored Composer and more splitting
