+++ extension/js/common/composer/interfaces/comopserr-errors.ts
@@ -1,0 +1,9 @@
+
+export type RecipientErrsMode = 'harshRecipientErrs' | 'gentleRecipientErrs';
+
+export class ComposerUserError extends Error { }
+export class ComposerNotReadyError extends ComposerUserError { }
+export class ComposerResetBtnTrigger extends Error { }
+
+export const PUBKEY_LOOKUP_RESULT_FAIL: 'fail' = 'fail';
+export const PUBKEY_LOOKUP_RESULT_WRONG: 'wrong' = 'wrong';
